requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
idna==3.7
ijson==3.3.0
psycopg==3.2.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
